+++ src/annotations/core/imageEffects/InvertColorImageEffect.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Damir Porobic <damir.porobic@gmx.com>
+ * Copyright (C) 2021 Damir Porobic <damir.porobic@gmx.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,20 +17,26 @@
  * Boston, MA 02110-1301, USA.
  */
 
-#include "ImageEffectFactory.h"
+#ifndef KIMAGEANNOTATOR_INVERTCOLORIMAGEEFFECT_H
+#define KIMAGEANNOTATOR_INVERTCOLORIMAGEEFFECT_H
 
-QGraphicsEffect *kImageAnnotator::ImageEffectFactory::create(kImageAnnotator::ImageEffects effect)
+#include <QGraphicsEffect>
+#include <QPainter>
+
+namespace kImageAnnotator {
+
+class InvertColorImageEffect : public QGraphicsEffect
 {
-	switch (effect) {
-		case ImageEffects::DropShadow:
-			return new DropShadowImageEffect;
-		case ImageEffects::Border:
-			return new BorderImageEffect;
-		case ImageEffects::Grayscale:
-			return new GrayscaleImageEffect;
-		case ImageEffects::InvertColor:
-			return new InvertColorImageEffect;
-		default:
-			return new NoImageEffect;
-	}
-}
+Q_OBJECT
+public:
+	explicit InvertColorImageEffect() = default;
+	~InvertColorImageEffect() override = default;
+
+protected:
+	void draw(QPainter * painter) override;
+};
+
+} // namespace kImageAnnotator
+
+
+#endif //KIMAGEANNOTATOR_INVERTCOLORIMAGEEFFECT_H